localize assets, better paths for client caching, head & scripts views
