--- assets/main.css
@@ -1,156 +1,0 @@
-/*
- * Globals
- */
-
-/* Links */
-a,
-a:focus,
-a:hover {
-    color: #fff;
-}
-
-/*
- * Base structure
- */
-
-html,
-body {
-    background-color: #181A1B;
-}
-
-body {
-    display: -ms-flexbox;
-    display: flex;
-    color: #fff;
-    text-shadow: 0 .05rem .1rem rgba(0, 0, 0, .5);
-    box-shadow: inset 0 0 5rem rgba(0, 0, 0, .5);
-}
-
-.cover-container {
-    max-width: 42em;
-}
-
-
-/*
- * Header
- */
-.masthead {
-    margin-bottom: 2rem;
-}
-
-.masthead-brand {
-    margin-bottom: 0;
-}
-
-.nav-masthead .nav-link {
-    padding: .25rem 0;
-    font-weight: 700;
-    color: rgba(255, 255, 255, .5);
-    background-color: transparent;
-    border-bottom: .25rem solid transparent;
-}
-
-.nav-masthead .nav-link:hover,
-.nav-masthead .nav-link:focus {
-    border-bottom-color: rgba(255, 255, 255, .25);
-}
-
-.nav-masthead .nav-link+.nav-link {
-    margin-left: 1rem;
-}
-
-.nav-masthead .active {
-    color: #fff;
-    border-bottom-color: #fff;
-}
-
-@media (min-width: 48em) {
-    .masthead-brand {
-        float: left;
-    }
-
-    .nav-masthead {
-        float: right;
-    }
-}
-
-
-/*
- * Cover
- */
-.cover {
-    padding: 0 1.5rem;
-}
-
-.cover .btn-lg {
-    padding: .75rem 1.25rem;
-    font-weight: 700;
-}
-
-
-/*
- * Footer
- */
-.mastfoot {
-    color: rgba(255, 255, 255, .5);
-}
-
-.jumbotron {
-    background-color: #1d1f21;
-}
-
-.card {
-    background-color: #1d1f21;
-}
-
-.text-muted {
-    color: #c1c1c1 !important;
-}
-
-.popover-header {
-    background-color: #7289DA;
-    color: #fff;
-}
-
-.popover-body {
-    background-color: #282a2e;
-    color: #fff;
-}
-
-.roundedbox {
-    border-radius: 40px;
-    background: #1d1f21;
-    padding: 20px;
-    width: 300px;
-    height: 100px;
-    text-align: center;
-    margin: 0 auto;
-}
-
-.dropdown-menu {
-    background-color: #282a2e;
-}
-
-.dropdown .dropdown-menu a {
-    color: #fff;
-    text-decoration: none;
-    background-color: #282a2e;
-}
-
-.dropdown-menu .dropdown-item {
-    display: inline-block;
-    width: 100%;
-    padding: 10px 5px;
-}
-
-.container .dropdown .dropdown-menu a:hover {
-    color: #fff;
-    background-color: #7489d6;
-    border-color: #fff;
-}
-
-.dropdown-item.active {
-    color: #fff;
-    background-color: #7489d6;
-    border-color: #fff;
-}
