+++ includes/views/scripts.php
@@ -1,0 +1,23 @@
+<script src="/static_assets/vendor/js/jquery.min.js"></script>
+<script src="/static_assets/vendor/js/bootstrap.bundle.min.js"></script>
+<script>
+    window.dataLayer = window.dataLayer || [];
+
+    function gtag() {
+        dataLayer.push(arguments);
+    }
+    gtag('js', new Date());
+
+    gtag('config', 'UA-119547589-1');
+
+    $(function() {
+        $('[data-toggle="tooltip"]').tooltip()
+    })
+</script>
+<script src="https://cdn.jsdelivr.net/npm/@widgetbot/crate@3" async defer>
+    new Crate({
+        server: '571922612230422528',
+        channel: '571922612230422530',
+        shard: 'https://disweb.deploys.io'
+    })
+</script>
